feat(tmux): named resurrect sessions with selection interface

+++ utils/tmux/resurrect-restore.sh
@@ -4,8 +4,21 @@ session_file=$1
 
 tmux_ressurect_sessions_path="$HOME/.local/share/tmux/resurrect"
 
-last_session_file="$tmux_ressurect_sessions_path/last"
+restore_script=$(tmux show-options -gqv @resurrect-restore-script-path)
 
-[[ -L $last_session_file ]] && rm $last_session_file
+if [[ -n $restore_script ]]; then 
 
-ln -s $session_file $last_session_file
+    last_session_file="$tmux_ressurect_sessions_path/last"
+
+    [[ -L $last_session_file ]] && rm $last_session_file
+
+    ln -s $session_file $last_session_file
+
+    tmux run $restore_script
+
+else
+
+    tmux display-message -d 3 "Resurrection script not found"
+    exit 1
+
+fi
